feat: job update event worker
