+++ apps/event-worker/src/workers/job-update/job-metadata.ts
@@ -1,0 +1,48 @@
+import type { Tx } from "@ctrlplane/db";
+
+import { sql } from "@ctrlplane/db";
+import * as schema from "@ctrlplane/db/schema";
+import { ReservedMetadataKey } from "@ctrlplane/validators/conditions";
+
+export const updateJobMetadata = async (
+  db: Tx,
+  jobId: string,
+  existingMetadata: schema.JobMetadata[],
+  metadata: Record<string, any>,
+) => {
+  const { [ReservedMetadataKey.Links]: links, ...remainingMetadata } = metadata;
+
+  if (links != null) {
+    const updatedLinks = JSON.stringify({
+      ...JSON.parse(
+        existingMetadata.find(
+          (m) => m.key === String(ReservedMetadataKey.Links),
+        )?.value ?? "{}",
+      ),
+      ...links,
+    });
+
+    await db
+      .insert(schema.jobMetadata)
+      .values({ jobId, key: ReservedMetadataKey.Links, value: updatedLinks })
+      .onConflictDoUpdate({
+        target: [schema.jobMetadata.jobId, schema.jobMetadata.key],
+        set: { value: updatedLinks },
+      });
+  }
+
+  if (Object.keys(remainingMetadata).length > 0)
+    await db
+      .insert(schema.jobMetadata)
+      .values(
+        Object.entries(remainingMetadata).map(([key, value]) => ({
+          jobId,
+          key,
+          value: JSON.stringify(value),
+        })),
+      )
+      .onConflictDoUpdate({
+        target: [schema.jobMetadata.jobId, schema.jobMetadata.key],
+        set: { value: sql`excluded.value` },
+      });
+};
