+++ apps/webservice/src/app/api/github/webhook/handlers/workflow_run.ts
@@ -93,9 +93,8 @@ export const handleWorkflowWebhookEvent = async (event: WorkflowRunEvent) => {
   const Run = `https://github.com/${repository.owner.login}/${repository.name}/actions/runs/${id}`;
   const Workflow = `${Run}/workflow`;
   await updateJob(
-    db,
     job.id,
-    { status, externalId, startedAt, completedAt },
-    { [String(ReservedMetadataKey.Links)]: { Run, Workflow } },
+    { status, externalId, startedAt, completedAt, updatedAt },
+    { [String(ReservedMetadataKey.Links)]: { Run, Workflow } } as any,
   );
 };
